Removed unused (and buggy) code from PARSE-MIME
RECURSIVE-P keyword argument to PARSE-MIME was never used (the original
intention of it might have been to give PARSE-MIME the whole content of HTTP
request).  To my knowledge this functionality is not used by anyone.

Additionally, this code contained a bug (pointed out by Thomas Bakketun):
every field was expected to have a content-type header, which of course is not
true for most [non-file] fields.