Changes to :write-content-to-file parameter of PARSE-MIME
Previously only values NIL and T were accepted, where T meant that
content of a MIME part is written to a file instead of returned as a
string.

This change also allows to give a function, which is called with MIME
part's headers, and should return either a pathname of a file where
the MIME part should be written, or an open stream [to a file] where
the MIME part is to be written.